please lint.
git-svn-id: file:///svn/unbound/trunk@3411 be551aaa-1e26-0410-a405-d3ace91eadb9

+++ services/cache/dns.c
@@ -682,7 +682,7 @@ fill_any(struct module_env* env,
 		/* create msg if none */
 		if(!msg) {
 			msg = dns_msg_create(qname, qnamelen, qtype, qclass,
-				region, num-i);
+				region, (size_t)(num-i));
 			if(!msg) {
 				lock_rw_unlock(&rrset->entry.lock);
 				return NULL;
